+++ include/NumCpp/Coordinates/ReferenceFrames/LLA.hpp
@@ -29,6 +29,7 @@
 
 #include <iostream>
 
+#include "NumCpp/NdArray.hpp"
 #include "NumCpp/Utils/essentiallyEqual.hpp"
 
 namespace nc::coordinates::reference_frames
@@ -60,6 +61,16 @@ namespace nc::coordinates::reference_frames
             longitude(inLongitude),
             altitude(inAltitude)
         {
+        }
+
+        /**
+         * @brief Convert to NdArray
+         *
+         */
+        [[nodiscard]] NdArray<double> toNdArray() const
+        {
+            NdArray<double> returnArray = { latitude, longitude, altitude };
+            return returnArray.transpose();
         }
 
         /**