+++ docs/markdown/ReleaseNotes.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## Version 2.18.0
+
+* added `toNdarray` to `coordinates::Cartesian`
+* added `toNdarray` to `coordinates::Euler`
+* added `toNdarray` to `coordinates::Orientation`
+* added `toNdarray` to `coordinates::reference_frames::AER`
+* added `toNdarray` to `coordinates::reference_frames::Geocentric`
+* added `toNdarray` to `coordinates::reference_frames::LLA`
+
 ## Version 2.17.0
 
 * added angular rate propagation methods to `rotations::Quaternion`